python_tools/requirements.txt
# STM32N6 Face Recognition UI Requirements
# DearPyGui-based modern interface

# Core UI framework
dearpygui==1.11.1

# Computer vision and image processing
opencv-python>=4.8.0
numpy>=1.24.0

# Serial communication
pyserial>=3.5

# Logging and utilities
typing-extensions>=4.0.0

# Optional: For advanced features
# matplotlib>=3.7.0  # For plotting embeddings
# pillow>=10.0.0     # For image format support